fix loginPage configuration utilize registrationId rather than realm
The web security configuration gets realm property injected to construct a login page configuration.
However, the registration id is the base for the correct login page address, not the realm name.
The example worked by coincidence, since the realm name matched the registration id.
This commit fixes the configuration by renaming the registration id uniquely and changing login page configuration to use it.